+++ packages/voictents-and-estinants-engine/src/adapter/estinant-builder/assembler/estinantAssembler.ts
@@ -1,25 +1,25 @@
-import { RightInputAppreffinge } from '../../core/engine-shell/appreffinge/rightInputAppreffinge';
+import { RightInputAppreffinge } from '../../../core/engine-shell/appreffinge/rightInputAppreffinge';
 import {
   Estinant2,
   GenericEstinant2,
-} from '../../core/engine-shell/estinant/estinant';
-import { GenericTropoignant2 } from '../../core/engine-shell/estinant/tropoignant';
-import { GenericIndexedHubblepup } from '../../core/engine-shell/quirm/hubblepup';
-import { GenericLeftInputVicken } from '../../core/engine-shell/vicken/leftInputVicken';
+} from '../../../core/engine-shell/estinant/estinant';
+import { GenericTropoignant2 } from '../../../core/engine-shell/estinant/tropoignant';
+import { GenericIndexedHubblepup } from '../../../core/engine-shell/quirm/hubblepup';
+import { GenericLeftInputVicken } from '../../../core/engine-shell/vicken/leftInputVicken';
 import {
   GenericRightInputHubblepupTupleVicken,
   GenericRightInputVickenTuple,
   GenericRightInputVoictentVicken,
-} from '../../core/engine-shell/vicken/rightInputVicken';
-import { Zorn, ZornTuple } from '../../utilities/semantic-types/zorn';
+} from '../../../core/engine-shell/vicken/rightInputVicken';
+import { Zorn, ZornTuple } from '../../../utilities/semantic-types/zorn';
 import {
   AssemblerContext,
   CoreConstituentOutputEntry,
-} from './estinantBuilderContext';
+} from '../shared/estinantBuilderContext';
 import {
   CoreOutputVickenFromAdaptedOutputVickenTuple,
   GenericAdaptedOutputVickenTuple,
-} from './vicken';
+} from '../shared/vicken';
 
 type EstinantAssembler<
   TLeftInputVicken extends GenericLeftInputVicken,
