+++ packages/voictents-and-estinants-engine/src/adapter/estinant-builder/output/outputHubblepupConditionalAppreffingeBuilder.ts
@@ -1,21 +1,21 @@
-import { Hubblepup } from '../../core/engine-shell/quirm/hubblepup';
-import { GenericVoque } from '../../core/engine/voque';
-import { Predicate } from '../../utilities/predicate';
-import { SpreadN } from '../../utilities/spreadN';
+import { Hubblepup } from '../../../core/engine-shell/quirm/hubblepup';
+import { GenericVoque } from '../../../core/engine/voque';
+import { Predicate } from '../../../utilities/predicate';
+import { SpreadN } from '../../../utilities/spreadN';
 import {
   buildEstinantAssembler,
   EstinantAssemblerParent,
-} from './estinantAssembler';
+} from '../assembler/estinantAssembler';
 import {
   buildInputOutputContextFromConstituentResultNormalizer,
   ConstituentResultNormalizer,
   InputOutputContext,
-} from './estinantBuilderContext';
+} from '../shared/estinantBuilderContext';
 import {
   AdaptedOutputVicken,
   GenericAdaptedLeftInputHubblepupVicken,
   GenericAdaptedOutputVickenTuple,
-} from './vicken';
+} from '../shared/vicken';
 
 type EmptyAdaptedRightInputVickenTuple = [];
